Move logging config to config.py
